Risk/BACK/secureguard/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
distlib==0.3.9
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.16.1
importlib_resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.1
pipenv==2024.1.0
platformdirs==4.3.6
psycopg==3.2.3
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
setuptools==75.1.0
sqlparse==0.5.1
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
virtualenv==20.26.6
